requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pytest
python-dateutil>=2.8.2

